refactor: split collections and courseware_import code out of libraries api
